Move mixable_versioned_table to framework directory
