+++ jubatus/core/driver/wscript
@@ -7,7 +7,6 @@ def configure(conf):
 def build(bld):
   source = [
       'driver.cpp',
-      'mixable_versioned_table.cpp',
       'classifier.cpp',
       'regression.cpp',
       'recommender.cpp',
@@ -25,7 +24,6 @@ def build(bld):
       'stat.hpp',
       'anomaly.hpp',
       'graph.hpp',
-      'mixable_versioned_table.hpp',
       'nearest_neighbor.hpp',
       'clustering.hpp',
       ]
