+++ jubatus/core/framework/wscript
@@ -9,6 +9,7 @@ def build(bld):
   source = [
       'model.cpp',
       'mixable.cpp',
+      'mixable_versioned_table.cpp',
       'push_mixable.cpp',
       'linear_mixable.cpp',
       'linear_function_mixer.cpp',
@@ -19,6 +20,7 @@ def build(bld):
       'linear_function_mixer.hpp',
       'mixable.hpp',
       'mixable_helper.hpp',
+      'mixable_versioned_table.hpp',
       'model.hpp',
       'packer.hpp',
       'push_mixable.hpp',
